--- @types/webworker.d.ts
@@ -1,33 +1,0 @@
-type WebWorkerMessage = {
-  id:string
-  userId:string
-  userName:string
-  type:'player'
-  action:string
-  parameter:string | number
-}
-
-type WebParentMessage = {
-  id:string
-  error:string
-  status:PlayerStatus
-}
-
-type WebUser = {
-  status:'known'
-  discord: {
-    id: string
-    username: string
-    discriminator: string
-  }
-  spotify?: {
-		id:string,
-		username:string,
-		locale:string
-	}
-  napster?: {
-    id:string,
-    username:string,
-    locale:string
-  }
-}
