--- client/src/types.d.ts
@@ -1,10 +1,0 @@
-export type User = {
-	status: 'new' | 'known'
-	discord?: {
-    id: string,
-    username: string,
-    discriminator: string,
-}
-	spotify?: { username: string }
-	napster?: { username: string }
-}
